code extensibility

+++ src/test/scala/com/lprakashv/FloatingNumberTest.scala
@@ -5,12 +5,12 @@ import org.scalatest.FunSuite
 class FloatingNumberTest extends FunSuite {
   test("1/3 == 0.333") {
     import com.lprakashv.commons.FloatingNumber.Implicits._
-    assert(1.0 / 3 ~= 0.333)
+    assert(BigDecimal(1.0) / 3 ~= 0.333)
   }
 
   test("PI != 0.33 with 0.0001 precision") {
     import com.lprakashv.commons.FloatingNumber.Implicits._
-    assert(!(math.Pi ~= (3.1, 0.0001)))
+    assert(!(math.Pi ~= (BigDecimal(3.1), 0.0001)))
   }
 
   test("PI == 0.33 with 0.1 precision") {
@@ -20,6 +20,6 @@ class FloatingNumberTest extends FunSuite {
 
   test("PI == 0.33") {
     import com.lprakashv.commons.FloatingNumber.Implicits._
-    assert(math.Pi ~= (3.1, 0.1))
+    assert(math.Pi.floatValue() ~= (3.1f, 0.1))
   }
 }
